+BigDecimal TextViewHolder

+++ app/src/main/java/com/example/recyclerviewimpl/data/models/Item.kt
@@ -1,10 +1,8 @@
 package com.example.recyclerviewimpl.data.models
 
-import android.graphics.Bitmap
-
 data class Item(
     val order: Int,
     val img: String,
     val title: String,
-    val cost: Long
+    val cost: Double
 )
